ROKIA/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openpyxl>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7
google-generativeai>=0.8.0
pydantic>=2.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
